requirements.txt
kafka-python==2.0.2
faker==20.1.0
pyspark==3.5.0
psycopg2-binary==2.9.9
streamlit==1.29.0
pandas==2.1.4
python-dotenv==1.0.0